Move properties. Introduce trainer._fitting
